remove adapter dataclass
